Getting things to compile

+++ DemonCastle.Editor/Windows/Properties/FloatProperty.cs
@@ -17,7 +17,7 @@ namespace DemonCastle.Editor.Windows.Properties {
 			
 			AddChild(SpinBox = new SpinBox {
 				CustomMinimumSize = new Vector2(20, 20),
-				SizeFlagsHorizontal = (int)SizeFlags.ExpandFill,
+				SizeFlagsHorizontal = SizeFlags.ExpandFill,
 				Step = 0.01,
 				Rounded = false,
 				Value = Binding.Get()
